+++ packages/codemod/src/transforms/replace-color-design-token/__tests__/replace-vars-color.test.ts
@@ -1,0 +1,5 @@
+import { runFixtureTests } from "../../../utils/test.js";
+import { join } from "node:path";
+import transform from "../index.js";
+
+runFixtureTests("replace-color-design-token", transform, join(__dirname, "..", "__testfixtures__"));