+++ packages/codemod/src/transforms/replace-tailwind-color/__testfixtures__/bg.output.tsx
@@ -18,8 +18,10 @@ export function BackgroundExample() {
       <div className="bg-palette-carrot-700">Scale Carrot High Background</div>
       <div className="bg-palette-carrot-700">Scale Carrot High Background</div>
       <div className="bg-palette-carrot-800">Scale Carrot High Background</div>
+      <div className="!bg-palette-carrot-800">Scale Carrot High Background</div>
       <div className="bg-palette-static-black">Static Black Background</div>
       <div className="bg-palette-static-white">Static White Background</div>
+      <div className="bg-staticGray900">Static Gray900 Background</div>
     </div>)
   );
 }