+++ packages/codemod/src/transforms/replace-tailwind-typography/__tests__/replace-vars-typography.test.ts
@@ -2,4 +2,8 @@ import { runFixtureTests } from "../../../utils/test.js";
 import { join } from "node:path";
 import transform from "../index.js";
 
-runFixtureTests("replace-vars-color", transform, join(__dirname, "..", "__testfixtures__"));
+runFixtureTests(
+  "replace-tailwind-typography",
+  transform,
+  join(__dirname, "..", "__testfixtures__"),
+);