+++ packages/codemod/src/transforms/replace-text-component/__tests__/replace-text-component.test.ts
@@ -1,0 +1,5 @@
+import { join } from "node:path";
+import { runFixtureTests } from "../../../utils/test.js";
+import transform from "../index.js";
+
+runFixtureTests("replace-text-component", transform, join(__dirname, "..", "__testfixtures__"));