+++ packages/codemod/src/transforms/replace-typography-design-token/__testfixtures__/basic.input.tsx
@@ -1,0 +1,10 @@
+// @ts-nocheck
+
+import { classNames } from "@seed-design/design-token";
+
+const typography = {
+  one: classNames.$semantic.typography.title2Regular,
+  two: classNames.$semantic.typography.label3Regular,
+  three: classNames.$semantic.typography.label5Regular,
+  four: classNames.$semantic.typography.label6Regular,
+}