+++ packages/codemod/src/transforms/replace-typography-design-token/__testfixtures__/variable-declarator.input.tsx
@@ -1,0 +1,6 @@
+import { classNames } from "@seed-design/design-token";
+
+const label3Bold = classNames.$semantic.typography.label3Bold;
+const label3Regular = classNames.$semantic.typography.label3Regular;
+const label2Regular = classNames.$semantic.typography.label2Regular;
+const caption2Regular = classNames.$semantic.typography.caption2Regular;