+++ packages/codemod/src/transforms/replace-typography-design-token/__testfixtures__/variable-declarator.output.tsx
@@ -1,0 +1,6 @@
+import { text } from "@seed-design/css/recipes/text";
+
+const label3Bold = text({ textStyle: "t4Bold" });
+const label3Regular = text({ textStyle: "t4Regular" });
+const label2Regular = text({ textStyle: "t5Regular" });
+const caption2Regular = text({ textStyle: "t2Regular" });