+++ packages/codemod/src/transforms/replace-typography-design-token/__tests__/replace-typography-design-token.test.ts
@@ -1,0 +1,9 @@
+import { join } from "node:path";
+import { runFixtureTests } from "../../../utils/test.js";
+import transform from "../index.js";
+
+runFixtureTests(
+  "replace-typography-design-token",
+  transform,
+  join(__dirname, "..", "__testfixtures__"),
+);